+++ src/risclog/sqlalchemy/tests/test_fixtures.py
@@ -22,10 +22,3 @@ def example_model(test_model_factory):
 def test_fixtures__database__selenium_testing__1(database_1, example_model):
     """It keeps the session by default after commit."""
     assert example_model.foo == 'bar'
-
-
-def test_fixtures__database__selenium_testing__2(
-        database_1, database__selenium_testing, example_model):
-    """It removes the session if database__selenium_testing is used ."""
-    with pytest.raises(sqlalchemy.orm.exc.DetachedInstanceError):
-        example_model.foo == 'bar'